Rename index.js export, change function logic, improve result output
Updating index.js to remove use of seeded-mnemonic. Also renames the "generateIdentityBoundAccount" function to a shorter "identityBoundAccount" function, and modifies the parameters to accept an ordered true/false value for the "options". The formation of the returned result is simplified, and the option to return the entropy bits is replaced with returning the entropy as the hex hash of the entropy seed.